Crate thorin

Re-exports

pub extern crate object;

Structs

DwarfPackage

Builder for DWARF packages, add input objects/packages with add_input_object or input objects referenced by an executable with add_executable before accessing the completed object with finish.

Enums

Error

Diagnostics (and contexts) emitted during DWARF packaging.

MissingReferencedObjectBehaviour

Should missing DWARF objects referenced by executables be skipped or result in an error?

Traits

Session

Session is expected to be implemented by users of thorin, allowing users of thorin to decide how to manage data, rather than thorin having arenas internally.